Refactor
Refactored some of the script. Added descriptions of Challenges and Rewards to help with weighting.
Redid weighting. Things with 50 weight will have a chance at being selected now, things below 50 won't be selected at all. things with 0 weight shouldn't ever be selected, less than that are banished.
Added the rest of the reward types.
Put all coordinates at the top of the file so it's easier to change if yours are different.
Changed Reload button to F5 because well.. it's refresh for everything else...
Added manual mode (press F1 to check)
Added option to turn off clicking on options (for debugging)
Added option to turn off controller mode (so it doesn't click so often on the top right)